refactor: 비즈니스와 관련없는 테스트 코드 분리
