+++ src/test/java/calculator/StringLearningTest.java
@@ -1,0 +1,13 @@
+package calculator;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class StringLearningTest {
+    @Test
+    void temp() {
+        String[] split = "1".split(",");
+        assertThat(split).isEqualTo(new String[]{"1"});
+    }
+}
